Выполнил задание №5. Личный проект: время разбивать камни
