+++ src/components/button-show-more.js
@@ -1,0 +1,5 @@
+export const createButtonShowMoreTemplate = () => {
+  return (
+    `<button class="films-list__show-more">Show more</button>`
+  );
+};
